requirements.txt
google-api-python-client
google-auth-oauthlib
google-api-python-client
pandas
django-social-auth
social-auth-app-django
